Update README.md
# Timeout-Project-1
Time to practice dealing with asynchronous functions like setTimeout and setInterval. Your objective is to implement the functions in each file of the /problems directory. In addition to the prompts available at the top of each file, Mocha specs are provided to test your work.

To get started, use the following commands:

-cd into the project directory
-npm install to install any dependencies
-mocha to run the test cases